% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gtf.R
\name{download_gtf}
\alias{download_gtf}
\title{Download gene annotations
Download gene annotations in GTF file format}
\usage{
download_gtf(organism, version = 91)
}
\arguments{
\item{organism}{select orgnaism ('Homo_sapiens' or 'Mus_musculus' or 'Rattus_norvegicus')}

\item{version}{select gtf version (By default version = 91)}
}
\description{
Download gene annotations
Download gene annotations in GTF file format
}
