% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_center.R
\name{feature_center}
\alias{feature_center}
\title{Feature center eset}
\usage{
feature_center(object, feature_center, verbose = TRUE)
}
\arguments{
\item{object}{eSet}

\item{feature_center}{Vector of numerical indexes or \code{feature_ids} (row names) to be used for centering}

\item{verbose}{whether to report progress message (logical)}
}
\value{
feature centered eSet
}
\description{
Feature center eset
}
