% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{filter_features_nonzero_for_two_samples_in_some_subgroup}
\alias{filter_features_nonzero_for_two_samples_in_some_subgroup}
\title{Keep features that are non zero (and non-NA) for two samples in some subgroup}
\usage{
filter_features_nonzero_for_two_samples_in_some_subgroup(object)
}
\arguments{
\item{object}{eSet}
}
\value{
filtered eSet
}
\description{
Keep features that are non zero (and non-NA) for two samples in some subgroup
}
