% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{filter_features_nonzero_in_some_sample}
\alias{filter_features_nonzero_in_some_sample}
\title{Keep features that are non-zero (and non-NA) for some sample}
\usage{
filter_features_nonzero_in_some_sample(object)
}
\arguments{
\item{object}{eSet}
}
\value{
filtered eSet
}
\description{
Keep features that are non-zero (and non-NA) for some sample
}
