% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{is_available_for_some_feature}
\alias{is_available_for_some_feature}
\alias{filter_samples_available_for_some_feature}
\title{Filter samples available for some feature}
\usage{
is_available_for_some_feature(object)

filter_samples_available_for_some_feature(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
SummarizedExperiment
}
\description{
Filter samples available for some feature
}
