% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_signs_if_all_negative.R
\name{flip_sign_if_all_exprs_are_negative}
\alias{flip_sign_if_all_exprs_are_negative}
\title{Flip sign if all expr values are negative}
\usage{
flip_sign_if_all_exprs_are_negative(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}
}
\value{
updated object
}
\description{
Flip sign if all expr values are negative
}
