% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_median_ref_exprs.R
\name{get_geometric_median_exprs}
\alias{get_geometric_median_exprs}
\alias{get_arithmetic_median_exprs}
\alias{get_median_exprs}
\title{Get median (sample) exprs}
\usage{
get_geometric_median_exprs(object)

get_arithmetic_median_exprs(object)

get_median_exprs(object, method = "arithmetic")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{method}{'arithmetic' or 'geometric'}
}
\description{
Returns exprs of median sample
The arithmetic median computes the median across samples separately for each feature.
The geometric median corresponds to the point in nfeature-dimensional space for which
the total distance to all other points is minimal (See details).
}
\details{
The geometric median is robust against outliers and norm invariant.
It is also known as the spatial median or the L1 median.
We compute it using an implementation based on NLM, which
outperformed other implementations in accuracy and speed, and is
robust against NAS
\href{https://dx.doi.org/10.2139/ssrn.1690502}{Croux et al., 2010}
}
\examples{
require(magrittr)
if (require(subramanian.2016)){
   subramanian.2016::exiqon \%>\% get_median_exprs()
}
}
