% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_median_ref_exprs.R
\name{get_median_ref_exprs}
\alias{get_median_ref_exprs}
\title{Get median reference exprs}
\usage{
get_median_ref_exprs(object,
  ref = levels(autonomics.import::sdata(object)$subgroup)[1],
  method = "arithmetic")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ref}{reference subgroup level}

\item{method}{'arithmetic' or 'geometric'}
}
\value{
vector with median reference exprs
}
\description{
Get median reference exprs
}
