% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{guess_nondetect_representation}
\alias{guess_nondetect_representation}
\alias{guess_nondetect_representation.default}
\alias{guess_nondetect_representation.SummarizedExperiment}
\title{Guess representation of nondetects}
\usage{
guess_nondetect_representation(x, ...)

\method{guess_nondetect_representation}{default}(x, verbose = FALSE)

\method{guess_nondetect_representation}{SummarizedExperiment}(x,
  verbose = FALSE)
}
\arguments{
\item{x}{matrix or SummarizedExperiment}
}
\value{
NA or 0
}
\description{
Guess representation of nondetects
}
\examples{
require(magrittr)
if (require(graumann.lfq)){
   x <- graumann.lfq::lfq.intensities
   x \%>\% guess_nondetect_representation()
   x \%>\% autonomics.import::exprs() \%>\% guess_nondetect_representation()
}
}
