% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute}
\usage{
impute(object, random_seed = NULL)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{random_seed}{Integer used with \code{\link[R.utils]{withSeed}} for reproducible
imputation (the used \code{\link[imputeLCMD]{impute.QRILC}} implys random draws from a distribution)}
}
\value{
updated object
}
\description{
Data from proteomics and metabolomics LCMS experiments are missing not at random (MNAR).
They are well imputed with the QRILC (quantile regression imputation of left censored data).
}
\details{
See \href{https://cran.r-project.org/package=imputeLCMD}{imputeLCMD package on CRAN}
See \href{https://www.biorxiv.org/content/early/2017/08/17/171967}{Wei et al. 2017}
}
