% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{impute_common_nondetects}
\alias{impute_common_nondetects}
\alias{impute_consistent_nondetects}
\title{Impute consistent nondetects}
\usage{
impute_common_nondetects(object, imputefun = function(x)
  imputeLCMD::impute.QRILC(x)[[1]], verbose = FALSE)

impute_consistent_nondetects(object, imputefun = function(x)
  imputeLCMD::impute.QRILC(x)[[1]], svar = "subgroup")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{imputefun}{imputation function}

\item{verbose}{logical(1)}

\item{svar}{character(1)}
}
\value{
SummarizedExperiment with updated exprs
}
\description{
Impute values missing in all (subgroup) samples.
}
\examples{
require(magrittr)
if (require(graumann.lfq)){

   # Read object
   object <- system.file('extdata/proteinGroups.txt', package = 'graumann.lfq') \%>\%
             autonomics::read_proteingroups_asis() \%>\%
             autonomics::zero_to_na(verbose=TRUE) \%>\%
            (function(x){ x$subgroup <- autonomics.import::guess_subgroup_values(x); x}) \%>\%
             autonomics::log2transform(verbose=TRUE)

   # Common nondetects - missing in all samples
   object \%>\% autonomics.import::split_by_svar() \%>\%
              magrittr::extract2(2) \%>\%
              impute_common_nondetects()

   # Consistent nondetects - missing in all subgroup samples
   object \%>\% impute_consistent_nondetects()
}
}
