% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_ratios.R
\name{invert_collapsed_strings}
\alias{invert_collapsed_strings}
\title{Invert collapsed strings}
\usage{
invert_collapsed_strings(x, sep)
}
\arguments{
\item{x}{vector of collapsed strings}

\item{sep}{separator used to collapse strings}
}
\description{
Invert collapsed strings
}
\examples{
 invert_collapsed_strings(c('Ctrl_A', 'Ctrl_B'), '_')
}
