% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_ratios.R
\name{invert_ratios}
\alias{invert_ratios}
\title{Invert ratios for some subgroup levels}
\usage{
invert_ratios(object, invert_subgroups, subgroup_frac)
}
\arguments{
\item{object}{eset}

\item{invert_subgroups}{vector of subgroup levels that require inversion}

\item{subgroup_frac}{fractionator symbol for subgroup values}

\item{channel_frac}{fractionator symbol for channel values}
}
\value{
eset with inverted ratios for specified subgroups
}
\description{
Invert ratios for some subgroup levels
}
\examples{
require(magrittr)

# Stem cell comparison (MaxQuant)
#--------------------------------
if (require(autonomics.data)){
   file <- system.file('extdata/stemcomp/maxquant/proteinGroups.txt',
                        package = 'autonomics.data')
   object <- file \%>\%
             autonomics.import::load_proteingroups(infer_design_from_sampleids = TRUE)
   object \%>\% autonomics.import::sdata()

   object \%>\% autonomics.preprocess::invert_ratios(
                 invert_subgroups = c('E_EM', 'E_BM', 'EM_BM'),
                 subgroup_frac    = '_') \%>\%
              autonomics.import::sdata()
}
}
