% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{invnorm}
\alias{invnorm}
\title{Inverse normal transform samples}
\usage{
invnorm(object)
}
\arguments{
\item{object}{eset}
}
\value{
normalized eset
}
\description{
Inverse normal transform samples
}
