% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features.R
\name{is_available_in_all_samples}
\alias{is_available_in_all_samples}
\alias{filter_features_available_in_all_samples}
\title{Keep features that are available in all samples}
\usage{
is_available_in_all_samples(object)

filter_features_available_in_all_samples(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}
}
\value{
updated object
}
\description{
Keep features that are available in all samples
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.preprocess::is_available_in_all_samples()
   object \%>\% autonomics.preprocess::filter_features_available_in_all_samples()
}
}
