% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{is_common_nondetect}
\alias{is_common_nondetect}
\title{Is complete nondetect}
\usage{
is_common_nondetect(x, representation = guess_nondetect_representation(x))
}
\arguments{
\item{x}{matrix or SummarizedExperiment}

\item{representation}{nondetect representation}
}
\value{
logical matrix with same dims as x
}
\description{
Is feature nondetect in all samples
}
\examples{
require(magrittr)
if (require(graumann.lfq)){
   x <- graumann.lfq::lfq.intensities
   x \%>\% is_common_nondetect()
   x \%>\% autonomics.import::exprs() \%>\% is_common_nondetect()
}
}
