% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{is_nondetect}
\alias{is_nondetect}
\alias{is_nondetect.default}
\alias{is_nondetect.SummarizedExperiment}
\title{Is nondetect}
\usage{
is_nondetect(x, ...)

\method{is_nondetect}{default}(x,
  representation = guess_nondetect_representation(x))

\method{is_nondetect}{SummarizedExperiment}(x,
  representation = guess_nondetect_representation(x))
}
\arguments{
\item{x}{matrix or SummarizedExperiment}

\item{representation}{nondetect representation}
}
\value{
logical matrix with same dims as x
}
\description{
Is nondetect
}
\examples{
require(magrittr)
if (require(graumann.lfq)){
   x <- graumann.lfq::lfq.intensities
   x \%>\% is_nondetect()
   x \%>\% autonomics.import::exprs() \%>\% is_nondetect()
   x \%>\% autonomics.import::exprs() \%>\% magrittr::extract(, 1) \%>\% is_nondetect()
}
}
