% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor.R
\name{left_censor}
\alias{left_censor}
\alias{right_censor}
\title{Censor exprs}
\usage{
left_censor(object, threshold)

right_censor(object, threshold)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{threshold}{numeric scalar: threshold beyond which to censor values}
}
\value{
SummarizedExperiment
}
\description{
Censor exprs
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.preprocess::left_censor(13)
   object \%>\% autonomics.preprocess::right_censor(30)
}
}
