% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_sd.R
\name{mean_and_sd}
\alias{mean_and_sd}
\title{mean_and_sd}
\usage{
mean_and_sd(object, MARGIN = c(1, 2), by_subgroup = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{MARGIN}{a vector giving the subscripts which the calculations will
be done for. 1 indicates rows/features, 2 indicates columns/samples,
c(1, 2) indicates rows/features and columns/samples.}

\item{by_subgroup}{Whether or not to calculate summary statistics per subgroup
(using \code{autonomics.import::sdata(object)$subgroup}) - used for
\code{MARGIN == 1} only;}
}
\value{
updated object
}
\description{
Updates data with feature/sample (row/column)-wise mean,
standard deviation, as well as a completion fraction (number of \code{\link{NA}}/measurements.
Under the hood \code{\link[matrixStats]{rowMeans2}} and \code{\link[matrixStats]{rowSds}} or
the \code{col} equivalent are used, all with \code{na.rm = TRUE}.
}
\details{
For subgroups without replicates, \code{NA} is returned.
}
