% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_center.R
\name{median_center}
\alias{median_center}
\title{Median center each sample}
\usage{
median_center(object)
}
\arguments{
\item{object}{eSet}
}
\value{
median centered eSet
}
\description{
Median center each sample
}
