% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minusinf_to_na.R
\name{minusinf_to_na}
\alias{minusinf_to_na}
\title{Replace -inf by NA}
\usage{
minusinf_to_na(object)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}
}
\value{
updated object
}
\description{
Minus inf can be created as log(0). This function replaces such minus inf by NA
}
