% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_to_na_if_not_all_zero.R
\name{zero_to_na_if_not_all_zero}
\alias{zero_to_na_if_not_all_zero}
\alias{missify_inconsistent_zeroes}
\title{Missify inconsistent zeroes}
\usage{
zero_to_na_if_not_all_zero(no_zero = c("fail", "warn_passthrough",
  "passthrough")[1], ...)

missify_inconsistent_zeroes(object, verbose = FALSE)
}
\arguments{
\item{no_zero}{\code{\link{character}} defining how to react of there are no \code{0}s (zeros) present in the data set}

\item{object}{SummarizedExperiment}
}
\value{
updated object
}
\description{
Convert inconsistent zeroes into NAs. \cr
Inconsistent zeroes are those not replicated across samples of the same subgroup.
Such zeroes are likely caused by identification/quantification stochasticity.
Their proper representation is NA rather than zero.
Failure to make this distinction diminishes statistical power.
}
\details{
Inconsistent zeroes are common in LCMS proteomics and metabolomics, due to the
stochastic nature of the identification process.
}
