% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_center.R
\name{mode_center}
\alias{mode_center}
\title{Mode center eset}
\usage{
mode_center(object, verbose = TRUE)
}
\arguments{
\item{object}{eSet}

\item{verbose}{whether to report progress message (logical)}
}
\value{
mode centered eSet
}
\description{
Mode center eset
}
