% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_exprs_weakly_gt.R
\name{na_exprs_weakly_gt}
\alias{na_exprs_weakly_gt}
\alias{na_exprs_weakly_lt}
\title{NA exprs (weakly) beyond threshold}
\usage{
na_exprs_weakly_gt(object, threshold)

na_exprs_weakly_lt(object, threshold)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{threshold}{numeric(1)}
}
\value{
Updated SummarizedExperiment
}
\description{
NA exprs (weakly) beyond threshold
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.preprocess::na_exprs_weakly_gt(30)
   object \%>\% autonomics.preprocess::na_exprs_weakly_lt(13)
}
}
