% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minusinf_to_na.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Convert NAs to zeroes}
\usage{
na_to_zero(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
Updated SummarizedExperiment
}
\description{
Convert NAs to zeroes
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% na_to_zero()
   object \%>\% na_to_zero(condition)
}
}
