% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{normalize_samples_on_svar}
\alias{normalize_samples_on_svar}
\title{Normalize samples on svar}
\usage{
normalize_samples_on_svar(object, normvar, result_dir = NULL)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{normvar}{svar on which to normalize}

\item{result_dir}{directory where normalization plots should be printed}
}
\value{
normalized SummarizedExperiment
}
\description{
Normalize samples on svar
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase
   object \%>\% autonomics.preprocess::normalize_samples_on_svar('PROTEIN_CONTENT')
}
}
