% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{normalize_samples_to_standard_normal}
\alias{normalize_samples_to_standard_normal}
\title{Normalize samples to standard normal}
\usage{
normalize_samples_to_standard_normal(object)
}
\arguments{
\item{object}{eset}
}
\description{
Normalize samples to standard normal
}
