% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{qrilc_common_nondetects}
\alias{qrilc_common_nondetects}
\title{QRILC common nondetects}
\usage{
qrilc_common_nondetects(object,
  representation = guess_nondetect_representation(object), verbose = FALSE)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{representation}{representation of missing values. Currently supported are NA, 0, and -Inf}

\item{verbose}{logical(1)}
}
\value{
SummarizedExperiment with updated exprs
}
\description{
QRILC-impute values missing in all samples.
}
\examples{
if require(graumann.lfq){
   object <- graumann.lfq::lfq.intensities \%>\%
             autonomics.import::split_by_svar() \%>\%
             magrittr::extract2(2)
   object \%>\% qrilc_common_nondetects()
}
}
