% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrilc_consistent_nondetects.R
\name{qrilc_consistent_nondetects}
\alias{qrilc_consistent_nondetects}
\title{QRILC consistent nondetects}
\usage{
qrilc_consistent_nondetects(object, svar = "subgroup")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{svar}{character(1)}
}
\description{
QRILC nondetects shared across all samples of the same subgroup.
}
\examples{
require(magrittr)
if (require(graumann.lfq)){
   object <- graumann.lfq::lfq.intensities
   object \%>\% autonomics.preprocess::qrilc_consistent_nondetects()
}
}
