% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{quantnorm_within_subgroups}
\alias{quantnorm_within_subgroups}
\title{Quantile normalize within subgroups}
\usage{
quantnorm_within_subgroups(object)
}
\arguments{
\item{object}{eset}
}
\value{
normalized eset
}
\description{
Quantile normalize within subgroups
}
