% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_median_ref_exprs.R
\name{subtract_exprs}
\alias{subtract_exprs}
\title{Subtract values from exprs}
\usage{
subtract_exprs(object, values)
}
\arguments{
\item{object}{SummarizedExperiment}

\item{values}{values}
}
\description{
Subtract values from exprs
}
