% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_median_ref_exprs.R
\name{subtract_median_ref_exprs}
\alias{subtract_median_ref_exprs}
\title{Subtract median ref exprs}
\usage{
subtract_median_ref_exprs(object,
  ref = levels(autonomics.import::sdata(object)$subgroup)[1],
  method = "arithmetic")
}
\arguments{
\item{object}{SummarizedExperiment}

\item{ref}{reference subgroup level}

\item{method}{'arithmetic' or 'geometric'}
}
\value{
SummarizedExperiment
}
\description{
Subtract median ref exprs
}
\author{
Aditya Bhagwat
}
