% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{transform_to_normal}
\alias{transform_to_normal}
\title{Transform vector to normal distribution}
\usage{
transform_to_normal(x, mean, sd)
}
\arguments{
\item{x}{numeric vector}

\item{mean}{mean}

\item{sd}{standard deviation}
}
\value{
transformed vector
}
\description{
Transform vector to normal distribution
}
