% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_samples.R
\name{transform_to_standard_normal}
\alias{transform_to_standard_normal}
\title{Transform vector to standard normal distribution}
\usage{
transform_to_standard_normal(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
transformed vector
}
\description{
Transform vector to standard normal distribution
}
