% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_consistent_nas.R
\name{zero_to_na_if_not_all_zero}
\alias{zero_to_na_if_not_all_zero}
\alias{na_inconsistent_zeroes}
\alias{zero_consistent_nas}
\title{NA inconsistent zeroes. Zero consistent NAS.}
\usage{
zero_to_na_if_not_all_zero(no_zero = c("fail", "warn_passthrough",
  "passthrough")[1], ...)

na_inconsistent_zeroes(object, verbose = FALSE)

zero_consistent_nas(object, verbose = FALSE)
}
\arguments{
\item{no_zero}{\code{\link{character}} defining how to react of there are no \code{0}s (zeros) present in the data set}

\item{object}{SummarizedExperiment}
}
\value{
updated object

updated object
}
\description{
Convert consistent NAS into zeroes.
And convert inconsistent zeroes into NAS.
}
\details{
Inconsistent zeroes are those not replicated across samples of the same subgroup.
Such zeroes are likely caused by identification/quantification stochasticity.
Their proper representation is NA rather than zero.
Failure to make this distinction diminishes statistical power.

Inconsistent zeroes are common in LCMS proteomics and metabolomics, due to the
stochastic nature of the identification process.
}
\examples{
if (require(subramanian.2016)){
   object <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                          package = 'subramanian.2016') \%>\%
             autonomics.import::load_exiqon(infer_design_from_sampleids = TRUE)
}
}
