% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minusinf_to_na.R
\name{zero_to_na}
\alias{zero_to_na}
\title{Convert zeroes to NAs}
\usage{
zero_to_na(object)
}
\arguments{
\item{object}{SummarizedExperiment}
}
\value{
Updated SummarizedExperiment
}
\description{
Convert zeroes to NAs
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   object <- autonomics.data::glutaminase \%>\% na_to_zero()
   object \%>\% autonomics.preprocess::zero_to_na()
}
}
