% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute}
\usage{
impute(
  object,
  method = c("none", "missForest", "impute.minDet", "impute.minProb", "impute.QRILC"),
  random_seed = NULL,
  ...
)
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{method}{Imputation method employed. See Section 'Details'.}

\item{random_seed}{Integer used with \code{\link[R.utils]{withSeed}} for reproducible
imputation (the used \code{\link[imputeLCMD]{impute.QRILC}} implys random draws from a distribution)}

\item{...}{Parameters handed on to the imputation methods.}
}
\value{
updated object
}
\description{
Data from proteomics and metabolomics LCMS experiments are missing not at random (MNAR).
They are well imputed with the QRILC (quantile regression imputation of left censored data).
}
\details{
See \href{https://cran.r-project.org/package=imputeLCMD}{imputeLCMD package on CRAN}

The function under the performs no imutation if \code{method = 'none'}, uses
 \code{\link[missForest]{missForest}} if \code{method = 'missForest'} and the parameter-named
 methods from \pkg{imputeLCMD} in all other cases.
}
