% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_ratios.R
\name{invert_ratios}
\alias{invert_ratios}
\title{Invert ratios for some subgroup levels}
\usage{
invert_ratios(object, invert_subgroups, channel_frac, subgroup_frac)
}
\arguments{
\item{object}{eset}

\item{invert_subgroups}{vector of subgroup levels that require inversion}

\item{channel_frac}{fractionator symbol for channel values}

\item{subgroup_frac}{fractionator symbol for subgroup values}
}
\value{
eset with inverted ratios for specified subgroups
}
\description{
Invert ratios for some subgroup levels
}
\examples{
require(magrittr)
if (require(autonomics.data)){
   proteingroups_file <- system.file('extdata/billing2016/proteinGroups.txt',
                                      package = 'autonomics.data')
   object <- autonomics.import::load_proteingroups(proteingroups_file)
   invert_subgroups <- unique(object$subgroup)
   object$subgroup;  subgroup_frac <- '_'
   object$channel;   channel_frac <- '/'
   object \%>\% autonomics.preprocess::invert_ratios(invert_subgroups, channel_frac, subgroup_frac)
}
}
