% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_consistent_nas.R
\name{zero_to_na_if_not_all_zero}
\alias{zero_to_na_if_not_all_zero}
\alias{na_inconsistent_zeroes}
\alias{zero_consistent_nas}
\title{Zero consistent NAs - NA inconsistent zeroes}
\usage{
zero_to_na_if_not_all_zero(no_zero = c("fail", "warn_passthrough",
  "passthrough")[1], ...)

na_inconsistent_zeroes(object, verbose = FALSE)

zero_consistent_nas(object, verbose = FALSE)
}
\arguments{
\item{no_zero}{\code{\link{character}} defining how to react of there are no \code{0}s (zeros) present in the data set}

\item{object}{SummarizedExperiment}
}
\value{
updated object

updated object
}
\description{
These functions allow to differentiate between consistent and inconsistent non-detects.
Inconsistent non-detects are due to measurement stochasticity.     These should be coded as NA rather than zero.
Consistent   non-detects are not due to measurement stochasticity. These should be coded as zero rather than NA.
Making this distinction improves power of statistical analysis.
}
\examples{
if (require(subramanian.2016)){
   object <- system.file('extdata/exiqon/subramanian.2016.exiqon.xlsx',
                          package = 'subramanian.2016') \%>\%
             autonomics.import::load_exiqon(infer_design_from_sampleids = TRUE)
}
}
