% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_to_na_if_not_all_zero.R
\name{zero_to_na_if_not_all_zero}
\alias{zero_to_na_if_not_all_zero}
\title{Zero to NA}
\usage{
zero_to_na_if_not_all_zero(object, no_zero = c("fail", "warn_passthrough",
  "passthrough")[1])
}
\arguments{
\item{object}{SummarizedExperiment, eSet, or EList}

\item{no_zero}{\code{\link{character}} defining how to react of there are no \code{0}s (zeros) present in the data set}
}
\value{
updated object
}
\description{
Prominently due to dynamic range limitations in mass spectrometry, data from proteomics
and metabolomics LCMS experiments is plagued by missing/dropout values, which occasionaly
(e.g. by MaxQuant LFQ) are reported as \code{0} (zero).
}
\details{
Based on the assumption that features truly absent will produce \code{0} (zero) measurements
across replicate samples, this function evaluates such data and replcaces \code{0} (zero)
with \code{NA} (not available) \bold{only} if any other replicate for the same feature is
\bold{not} \code{0} (zero) within the same subgroup.
}
